style: apply ruff --fix across src/ + tests/, split vivi kb_search import
CI was still failing because the lint config landed in pyproject.toml but
the matching autofixes (import sorting, dead import removal, blank-line
normalisation) were only present on disk locally — they were never
committed. CI ran ruff against the un-fixed committed tree and still hit
~500 style nits.

Changes
- Committed the ruff --fix output across the whole tree (133 files,
  mostly I001 import reordering + a handful of F401 removals).
- src/vivi.py: split the mechanic import onto two lines so each carries
  its own `# noqa: F401`. The previous combined form
  `from mechanic import get_advice_for_alert, search as kb_search`
  got the `kb_search` alias stripped by --fix because the noqa only
  applied to the first symbol — which broke the `vivi.kb_search` test
  seam again. Per-line noqa survives --unsafe-fixes.

Verification
- ruff check src tests → all checks passed
- ruff check --fix --unsafe-fixes src/vivi.py → kb_search survives
- pytest tests/ → 891 passed, 2 failed (Windows-only: 0o600 chmod,
  backslash path separator). CI runs on Linux.

+++ tests/test_realdash_bridge.py
@@ -7,10 +7,10 @@ UNCAGED TECHNOLOGY — EST 1991
 """
 
 import struct
+
 import pytest
 
 import realdash_bridge as rb
-
 
 REALDASH_HEADER = bytes([0x44, 0x33, 0x22, 0x11])
 
@@ -151,7 +151,7 @@ class TestRPMEncoding:
         rb.latest['coolant'] = 90.0
         frame = rb.pack_engine_frame()
         coolant_raw = struct.unpack_from('>h', frame, 10)[0]
-        assert coolant_raw == int((90 + 40) * 10)
+        assert coolant_raw == ((90 + 40) * 10)
 
 
 class TestExtraEngineEncoding:
